// src/app/interfaces/forms/register-form-colaboradores.ts
export interface RegisterFormColaboradores {
  identificationTypeId: number;
  identification: number;
  name1: string;
  name2: string;
  lastName1: string;
  lastName2: string;
  email: string;
  phoneNumber: string;
  mobile: string;
  politicId: number;
  authorizationId: number;
  captchaToken: string;
}